src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.74",
    "celery>=5.6.2",
    "django>=6.0.3",
    "django-celery-results>=2.6.0",
    "django-cors-headers>=4.9.0",
    "django-minio-backend>=4.5.0",
    "django-minio-storage",
    "django-private-storage>=3.1.3",
    "django-storages>=1.14.6",
    "djangorestframework>=3.17.0",
    "flower>=2.0.1",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-text-splitters>=1.1.1",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pypdf>=6.9.2",
    "redis>=7.3.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
]
